requirements.txt
###
# This file is only here to make sure that something like
#
#    pip install -e .
#
# works as expected. Requirements can be found in setup.py.
###

smbus2
numpy
joblib
octoprint
flask
RPi.GPIO
scikit-learn
